+++ linking-platform/link-picker/src/ui/__tests__/__helpers/index.ts
@@ -1,6 +1,5 @@
 import { LinkSearchListItemData } from '../../../../src/types';
 
-export { default as MockLinkPickerPlugin } from './plugin-mock';
 export { default as ManualPromise } from './manual-promise';
 
 export const getDefaultItems = (