+++ linking-platform/link-provider/src/ff.ts
@@ -1,0 +1,8 @@
+import { LinkingPlatformFeatureFlags } from '@atlaskit/linking-common';
+import { useSmartLinkContext } from './state/context';
+
+export function useFeatureFlag(featureFlag: keyof LinkingPlatformFeatureFlags) {
+  const { featureFlags } = useSmartLinkContext();
+
+  return featureFlags && featureFlags[featureFlag];
+}