+++ linking-platform/linking-common/src/ff.ts
@@ -1,0 +1,3 @@
+export interface LinkingPlatformFeatureFlags {
+  showHoverPreview?: boolean;
+}