+++ media/smart-card/docs/utils/props-element-item.tsx
@@ -3,4 +3,4 @@ import { ElementItem } from '../../src/view/FlexibleCard/components/blocks/types
 /**
  * Hack to display Flexible Smart Links `ElementItem` for documentation purpose.
  */
-export default (props: ElementItem) => null;
+export default (props: { props: ElementItem }) => null;