+++ linking-platform/link-test-helpers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/link-test-helpers
 
+## 1.1.0
+
+### Minor Changes
+
+- [`854d5c55d9b`](https://bitbucket.org/atlassian/atlassian-frontend/commits/854d5c55d9b) - Adds link-picker export with mock link picker plugin.
+
 ## 1.0.0
 
 ### Major Changes